Refactor move bookmark functionality
